+++ CHANGELOG.md
@@ -4,4 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.0.0 - 2021-04-06
+### Added
+- Initial version
+- Add command to load all fixtures
+- Add FixtureHelper with following methods:
+  - `getEuroCurrencyId`
+  - `getInvoicePaymentMethodId`
+  - `getNotSpecifiedSalutationId`
+  - `getGermanCountryId`